% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{api_query}
\alias{api_query}
\title{Perform an API Query Using httr2}
\usage{
api_query(base_url, path, parameters = NULL)
}
\arguments{
\item{base_url}{A string specifying the base URL of the API.}

\item{path}{A string specifying the path of the specific API endpoint.}

\item{parameters}{An optional named list of query parameters to be included
in the request.}
}
\value{
A list containing the following elements:
\itemize{
  \item{response}{The complete response object returned by the API.}
  \item{is_error}{Logical value indicating if the response was an error.}
  \item{status}{The HTTP status code of the response.}
  \item{description}{The description of the HTTP status.}
}
}
\description{
This function sends an HTTP request to a specified API endpoint and returns
various details about the response. It is designed to work with the httr2
package.
}
\details{
The function constructs a request using the base URL and the path.
If provided, query parameters are appended to the request. The function then
performs the request and checks for errors. It returns a list containing the
response, error status, HTTP status code, and the description of the status.
}
