% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{assert_columns}
\alias{assert_columns}
\title{Check if a tibble contains all required columns}
\usage{
assert_columns(tb, cols)
}
\arguments{
\item{tb}{Tibble or data frame to examine}

\item{cols}{A string vector with required column names}
}
\value{
TRUE if assertion passed
}
\description{
Check if a tibble contains all required columns
}
