% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{assert_http_file}
\alias{assert_http_file}
\title{Check if HTTP file exists}
\usage{
assert_http_file(url_file)
}
\arguments{
\item{url_file}{HTTP address of the file}
}
\value{
Nothing
}
\description{
Stops with error message if HTTP file is not accessible.
}
