% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{assert_species}
\alias{assert_species}
\title{Check if species are valid}
\usage{
assert_species(species, fetch_fun)
}
\arguments{
\item{species}{A string, species designation for a given database}

\item{fetch_fun}{A string, name of the function to retrieve available
species, must return a tibble with a column \code{designation}.}
}
\value{
A tibble with valid species - a response from \code{fetch_fun}
}
\description{
Check if species are valid
}
