% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{catch_error}
\alias{catch_error}
\title{Catch and Handle Errors Based on Specified Action}
\usage{
catch_error(server, resp, on_error = c("stop", "warn"))
}
\arguments{
\item{server}{Name of the server.}

\item{resp}{Response from the server.}

\item{on_error}{A character vector specifying the error handling method. It
can take values `"stop"` or `"warn"`. The default is `"stop"`. `"stop"`
will halt the function execution and throw an error, while `"warn"` will
issue a warning and return `NULL`.}
}
\value{
In case of `"warn"`, the function returns `NULL`. If `"stop"` is
  chosen, the function halts with an error and does not return a value.
}
\description{
This function handles errors by either stopping execution or issuing a
warning, depending on the specified action. It takes a message and an action
choice (`"stop"` or `"warn"`) as inputs.
}
