% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteractive_example.R
\name{enrichment_table}
\alias{enrichment_table}
\title{Helper function for Shiny app}
\usage{
enrichment_table(de, term_data, input, max_points = 10000)
}
\arguments{
\item{de}{Differential expression results}

\item{term_data}{All functional term data}

\item{input}{App input}

\item{max_points}{Maximum number of points allowed to select}
}
\value{
A tibble with functional enrichment results
}
\description{
Helper function for Shiny app
}
