% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exmpl_all}
\alias{exmpl_all}
\title{Example set of background genes.}
\format{
A character vector with 7075 elements.
}
\usage{
exmpl_all
}
\description{
A set of gene names for proteins that could be detected in a typical
proteomics experiment on human samples.
}
\keyword{datasets}
