% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exmpl_sel}
\alias{exmpl_sel}
\title{Example set of selected genes.}
\format{
A character vector with 21 elements.
}
\usage{
data(exmpl_sel)
}
\value{
A vector with selected gene names.
}
\description{
A set of gene names manually selected to illustrate functional enrichment.
}
\keyword{datasets}
