% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{extract_key}
\alias{extract_key}
\title{Helper function for `parse_wiki_gpml`}
\usage{
extract_key(s, key)
}
\arguments{
\item{s}{Input string}

\item{key}{Key to extract}
}
\value{
Value of the key
}
\description{
Helper function for `parse_wiki_gpml`
}
