% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{extract_obo_values}
\alias{extract_obo_values}
\title{Parse fields in obo data}
\usage{
extract_obo_values(obo, key)
}
\arguments{
\item{obo}{A character vector containing obo data}

\item{key}{A key to extract, e.g. "id" or "name"}
}
\value{
All values for the given key
}
\description{
Parse fields in obo data
}
