% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fenr-package.R
\docType{package}
\name{fenr-package}
\alias{fenr}
\alias{fenr-package}
\title{fenr: Fast functional enrichment for interactive applications}
\value{
Package fenr.
}
\description{
This R package provides a fast and efficient method for functional enrichment
analysis, optimized for speed and designed for use in interactive
applications, such as *Shiny* apps.

To learn more about fenr, start with the vignette: vignette("fenr").
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bartongroup/fenr}
  \item Report bugs at \url{https://github.com/bartongroup/fenr/issues}
}

}
\author{
\strong{Maintainer}: Marek Gierlinski \email{M.Gierlinski@dundee.ac.uk} (\href{https://orcid.org/0000-0001-9149-3514}{ORCID})

Authors:
\itemize{
  \item Marek Gierlinski \email{M.Gierlinski@dundee.ac.uk} (\href{https://orcid.org/0000-0001-9149-3514}{ORCID})
}

}
\keyword{internal}
