% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioplanet.R
\name{fetch_bp}
\alias{fetch_bp}
\title{Get functional term data from BioPlanet}
\usage{
fetch_bp()
}
\value{
A list with \code{terms} and \code{mapping} tibbles.
}
\description{
Get functional term data from BioPlanet
}
\examples{
\dontrun{
bioplanet_data <- fetch_bp()
}
}
