% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{fetch_go_from_bm}
\alias{fetch_go_from_bm}
\title{Get functional term data from Ensembl}
\usage{
fetch_go_from_bm(mart)
}
\arguments{
\item{mart}{Object class \code{Mart} representing connection to BioMart
database, created with, e.g., \code{useEnsembl}.}
}
\value{
A list with \code{terms} and \code{mapping} tibbles.
}
\description{
Download term information (GO term ID and name) and gene-term mapping
(gene ID, symbol and GO term ID) from Ensembl.
}
\examples{
\donttest{
mart <- biomaRt::useEnsembl(biomart = "ensembl", dataset = "scerevisiae_gene_ensembl")
go_terms <- fetch_go_from_bm(mart)
}
}
