% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{fetch_go_from_go}
\alias{fetch_go_from_go}
\title{Get functional term data from gene ontology}
\usage{
fetch_go_from_go(species)
}
\arguments{
\item{species}{Species designation. Base file name for species file under
\url{http://current.geneontology.org/annotations}. Examples are
\file{goa_human} for human, \file{mgi} for mouse or \file{sgd} for yeast.}
}
\value{
A list with \code{terms} and \code{mapping} tibbles.
}
\description{
Download term information (GO term ID and name) and gene-term mapping
(gene symbol and GO term ID) from gene ontology.
}
\examples{
go_data <- fetch_go_from_go("sgd")
}
