% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{fetch_go_genes_bm}
\alias{fetch_go_genes_bm}
\title{Download GO term gene mapping from Ensembl}
\usage{
fetch_go_genes_bm(mart)
}
\arguments{
\item{mart}{Object class \code{Mart} representing connection to BioMart
database, created with, e.g., \code{useEnsembl}.}
}
\value{
A tibble with ensembl_gene_is, gene_symbol and term_id.
}
\description{
Download GO term gene mapping from Ensembl
}
