% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{fetch_go_genes_go}
\alias{fetch_go_genes_go}
\title{Download GO term gene mapping from geneontology.org}
\usage{
fetch_go_genes_go(species)
}
\arguments{
\item{species}{Species designation. Base file name for species file under
\url{http://current.geneontology.org/annotations}. Examples are
\file{goa_human} for human, \file{mgi} for mouse or \file{sgd} for yeast.}
}
\value{
A tibble with columns \code{gene_symbol}, \code{uniprot_id} and \code{term_id}.
}
\description{
Download GO term gene mapping from geneontology.org
}
