% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{fetch_go_species}
\alias{fetch_go_species}
\title{Find all species available from geneontology.org}
\usage{
fetch_go_species(on_error = c("stop", "warn", "ignore"), all_sources = FALSE)
}
\arguments{
\item{on_error}{A character string indicating the error handling strategy:
either "stop" to halt execution, "warn" to issue a warning and return
`NULL` or "ignore" to return `NULL` without warnings. Defaults to "stop".}

\item{all_sources}{Logical, if TRUE, return all available annotation sources
for each species. If FALSE, return one preferred designation per species.}
}
\value{
A tibble with Gene Ontology species metadata. Column
  \code{designation} contains species designations used in function
  \code{fetch_go}.
}
\description{
This function downloads and parses Gene Ontology's organism metadata table
to return the current annotation file species designations. By default, one
preferred designation is returned for each species. Model Organism Database
(MOD) annotations are preferred when they are available; otherwise UniProt
annotations are returned. Set \code{all_sources = TRUE} to return both MOD
and UniProt designations.
}
\examples{
go_species <- fetch_go_species(on_error = "warn")
go_species_all <- fetch_go_species(on_error = "warn", all_sources = TRUE)
}
