% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{fetch_go_terms}
\alias{fetch_go_terms}
\title{Download GO term descriptions}
\usage{
fetch_go_terms(obo_file = "http://purl.obolibrary.org/obo/go.obo")
}
\arguments{
\item{obo_file}{A URL or a local file name containing GO ontology, in OBO format.}
}
\value{
A tibble with term_id and term_name.
}
\description{
Download GO term descriptions
}
