% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{fetch_kegg}
\alias{fetch_kegg}
\title{Get functional term data from KEGG}
\usage{
fetch_kegg(species)
}
\arguments{
\item{species}{KEGG species code, for example \code{"hsa"} for human. The
full list of available KEGG species can be found by using
\code{KEGGREST::keggList("organism")}. The column \code{organism} contains
the codes used here.}
}
\value{
A list with \code{terms} and \code{mapping} tibbles.
}
\description{
Download information (pathway ID and name) and gene-pathway mapping (entrez
gene ID, gene symbol and pathway ID) from KEGG. Gene symbols are extracted
from gene descriptions. For some species (e.g. yeast), gene symbols are
returned instead of entrez IDs and not in gene description. This function is
based on BioConductor package \code{KEGGREST}.
}
\examples{
\dontrun{
kegg_data <- fetch_kegg("hsa")
}
}
