% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{fetch_kegg_data}
\alias{fetch_kegg_data}
\title{A helper function to retrieve API data from KEGG}
\usage{
fetch_kegg_data(path)
}
\arguments{
\item{path}{Path in the query}
}
\value{
A string with result
}
\description{
A helper function to retrieve API data from KEGG
}
