% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{fetch_kegg_mapping}
\alias{fetch_kegg_mapping}
\title{Download term - gene mapping from KEGG}
\usage{
fetch_kegg_mapping(pathways, batch_size)
}
\arguments{
\item{pathways}{A character vector with KEGG pathways}

\item{batch_size}{Number of pathways sent to KEGG database in one query. The
maximum allowed is 10.}
}
\value{
A tibble with columns \code{gene_id} and \code{term_id}
}
\description{
Download term - gene mapping from KEGG
}
