% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{fetch_kegg_pathways}
\alias{fetch_kegg_pathways}
\title{Download pathway data from KEGG}
\usage{
fetch_kegg_pathways(species)
}
\arguments{
\item{species}{A valid species designation used by KEGG.}
}
\value{
A tibble with columns \code{gene_id} and \code{term_id}.
}
\description{
Download pathway data from KEGG
}
