% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{fetch_kegg_species}
\alias{fetch_kegg_species}
\title{Find all species available from KEGG}
\usage{
fetch_kegg_species()
}
\value{
A tibble, in which column \code{designation} contains species
  designations used in function \code{fetch_kegg}.
}
\description{
Find all species available from KEGG
}
\examples{
spe <- fetch_kegg_species()
}
