% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{fetch_reactome_data}
\alias{fetch_reactome_data}
\title{A helper function to retrieve API data from Reactome}
\usage{
fetch_reactome_data(path)
}
\arguments{
\item{path}{Path in the query}
}
\value{
A tibble with result
}
\description{
This is used to quickly retrieve Reactome species. Alas, Reactome does not
provide API to download all pathways, so we need to download them directly
from a URL.
}
