% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{fetch_reactome_ensembl_genes}
\alias{fetch_reactome_ensembl_genes}
\title{Download term - Ensembl gene ID mapping from Reactome}
\usage{
fetch_reactome_ensembl_genes(spec)
}
\arguments{
\item{spec}{Reactome species}
}
\value{
A tibble with columns \code{gene_id} and \code{term_id}
}
\description{
Download term - Ensembl gene ID mapping from Reactome
}
\details{
This function downloads one large file containing a mapping between
  Enxembl gene IDs and Reactome terms. This is significantly faster than
  \code{fetch_reactome_genes}, but if gene symbols are required, needs
  additional ID conversion.
}
