% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{fetch_reactome_gene_association}
\alias{fetch_reactome_gene_association}
\title{Download term - gene association from Reactome}
\usage{
fetch_reactome_gene_association(tax_id)
}
\arguments{
\item{tax_id}{Taxon ID}
}
\value{
A tibble with columns \code{accession_number}, \code{gene_symbol} and \code{term_id}
}
\description{
Download term - gene association from Reactome
}
