% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{fetch_reactome_pathways}
\alias{fetch_reactome_pathways}
\title{Download pathway data from Reactome}
\usage{
fetch_reactome_pathways(spec)
}
\arguments{
\item{spec}{Reactome species}
}
\value{
A tibble with columns \code{term_id} and \code{term_name}
}
\description{
Download pathway data from Reactome
}
