% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{fetch_reactome_species}
\alias{fetch_reactome_species}
\title{List of available Reactome species}
\usage{
fetch_reactome_species()
}
\value{
A character vector with species names used by Reactome.
}
\description{
List of available Reactome species
}
\examples{
re <- fetch_reactome_species()
}
