% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{fetch_wiki}
\alias{fetch_wiki}
\title{Get functional term data from WikiPathways}
\usage{
fetch_wiki(species)
}
\arguments{
\item{species}{WikiPathways species designation, for example "Homo
sapiens" for human. Full list of available species can be found using
\code{fetch_wiki_species()}.}
}
\value{
A list with \code{terms} and \code{mapping} tibbles.
}
\description{
Download term information (pathway ID and name) and gene-pathway mapping
(gene symbol and pathway ID) from WikiPathways.
}
\examples{
wiki_data <- fetch_wiki("Saccharomyces cerevisiae")
}
