% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{fetch_wiki_pathway_genes_api}
\alias{fetch_wiki_pathway_genes_api}
\title{Download WikiPathways gene-pathway mapping using API}
\usage{
fetch_wiki_pathway_genes_api(
  pathways,
  gene_dbs = c("Ensembl", "HGNC", "HGNC Accession number", "Uniprot-TrEMBL")
)
}
\arguments{
\item{pathways}{A character vector with pathway names.}

\item{gene_dbs}{Names of databases to use}
}
\value{
A tibble with \code{term_id} and \code{gene_symbol}
}
\description{
Download WikiPathways gene-pathway mapping using API
}
