% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{fetch_wiki_pathways}
\alias{fetch_wiki_pathways}
\title{Download pathway data from WikiPathways}
\usage{
fetch_wiki_pathways(species)
}
\arguments{
\item{species}{Species name recognised by WikiPathways (see \code{fetch_wiki_species()})}
}
\value{
A tibble with columns \code{gene_id} and \code{term_id}
}
\description{
Download pathway data from WikiPathways
}
