% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{fetch_wiki_species}
\alias{fetch_wiki_species}
\title{List of available WikiPathways species}
\usage{
fetch_wiki_species()
}
\value{
A character vector with species names used by WikiPathways.
}
\description{
List of available WikiPathways species
}
\examples{
spec <- fetch_wiki_species()
}
