% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_feature_term}
\alias{get_feature_term}
\title{Get terms annotating a given feature.}
\usage{
get_feature_term(term_data, feature_id)
}
\arguments{
\item{term_data}{An object class \code{fenr_terms}, created by
\code{prepare_for_enrichment}.}

\item{feature_id}{A string with a feature ID}
}
\value{
A character vector containing functional term IDs annotating given feature.
}
\description{
Get terms annotating a given feature.
}
