% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{get_wiki_query}
\alias{get_wiki_query}
\title{REST API query for WikiPathways}
\usage{
get_wiki_query(query, parameters = NULL)
}
\arguments{
\item{query}{Query, e.g. `listPathways`}

\item{parameters}{Parameters, e.g. `organism=Homo sapiens`}
}
\value{
JSON with response
}
\description{
REST API query for WikiPathways
}
