% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteractive_example.R
\name{get_xy_data}
\alias{get_xy_data}
\title{Helper function for Shiny app}
\usage{
get_xy_data(de, input)
}
\arguments{
\item{de}{Differential expression results}

\item{input}{App input}
}
\value{
A tibble with x, y
}
\description{
Helper function for Shiny app
}
