% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go}
\alias{go}
\title{GO-terms data downloaded for the vignette.}
\format{
A list of two tibbles
}
\usage{
data(go)
}
\value{
Contains GO-term descriptions and gene mapping.
}
\description{
Downloaded using \code{go <- fetch_go(species = "YEAST-mod")}
}
\keyword{datasets}
