% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_species}
\alias{go_species}
\title{GO species}
\format{
A tibble
}
\usage{
data(go_species)
}
\value{
Contains species available through Gene Ontology
}
\description{
Downloaded using \code{go_species <- fetch_go_species()}
}
\keyword{datasets}
