% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteractive_example.R
\name{main_plot}
\alias{main_plot}
\title{Helper function for Shiny app}
\usage{
main_plot(de, input)
}
\arguments{
\item{de}{Differential expression results}

\item{input}{App input}
}
\value{
Data for volcano or ma plot
}
\description{
Helper function for Shiny app
}
