% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{match_species}
\alias{match_species}
\title{Return a field for a given species}
\usage{
match_species(species, fetch_fun, col_name)
}
\arguments{
\item{species}{Species designation}

\item{fetch_fun}{A function to retrieve available species, must return a
tibble with a column \code{designation}.}

\item{col_name}{Column name in the tibble returned by \code{fetch_fun} to
extract, e.g. \code{tax_id}}
}
\value{
A value extracted from column \code{col_name} at row where
  \code{designation} = \code{species}.
}
\description{
Used to match species designation with a species ID.
}
