% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{parse_kegg_genes}
\alias{parse_kegg_genes}
\title{Parse flat file returned from KEGG GET API}
\usage{
parse_kegg_genes(s)
}
\arguments{
\item{s}{A string containing flat file result}
}
\value{
A tibble with \code{gene_is}, \code{gene_symbol} and \code{term_id}.
}
\description{
Parse flat file returned from KEGG GET API
}
