% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{parse_obo_file}
\alias{parse_obo_file}
\title{Parse OBO file and return a tibble with key and value}
\usage{
parse_obo_file(obo)
}
\arguments{
\item{obo}{Obo file content as a character vector}
}
\value{
A tibble with term_id, key and value
}
\description{
Parse OBO file and return a tibble with key and value
}
