% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{parse_wiki_gpml}
\alias{parse_wiki_gpml}
\title{Parse GPML string provided by WikiPathways}
\usage{
parse_wiki_gpml(gpml, keys = c("TextLabel", "Type", "Database", "ID"))
}
\arguments{
\item{gpml}{GMPL string returned by WikiPathways query.}

\item{keys}{Keys to be extracted.}
}
\value{
A tibble with selected keys and values
}
\description{
Parse GPML string provided by WikiPathways
}
