% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteractive_example.R
\name{plot_ma}
\alias{plot_ma}
\title{MA plot}
\usage{
plot_ma(d, fdr_limit = 0.05)
}
\arguments{
\item{d}{Tibble with x, y and FDR}

\item{fdr_limit}{FDR limit below which point will be plotted black}
}
\value{
A ggplot object
}
\description{
MA plot
}
