% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{remove_cache}
\alias{remove_cache}
\title{Remove all cache}
\usage{
remove_cache(ask = TRUE)
}
\arguments{
\item{ask}{Logical, whether to ask user for confirmation.}
}
\value{
TRUE if successfully removed.
}
\description{
This function will remove all cached data used by `fenr`. The user will be
prompted for confirmation. Use only when you suspect the cache was corrupted.
Use with caution!
}
\examples{
\dontrun{
remove_cache()
}
}
