% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{test_mapping}
\alias{test_mapping}
\title{Test if returned mapping is consistent with expected selection}
\usage{
test_mapping(returned, expected, feature_id)
}
\arguments{
\item{returned}{A tibble, the mapping returned by  fetch_*.}

\item{expected}{A tibble, a selection of expected mapping.}

\item{feature_id}{A character string, the name of the column containing the feature
ID.}
}
\description{
Test if returned mapping is consistent with expected selection
}
