% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{test_terms}
\alias{test_terms}
\title{Test if returned terms tibble is consistent with expected selection}
\usage{
test_terms(returned, expected)
}
\arguments{
\item{returned}{A tibble, the terms data returned by  fetch_*.}

\item{expected}{A tibble, a selection of expected terms data.}
}
\description{
Test if returned terms tibble is consistent with expected selection
}
