% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{assert_url_path}
\alias{assert_url_path}
\title{Check if URL with a path is responding OK}
\usage{
assert_url_path(url_path)
}
\arguments{
\item{url_path}{Full URL with a path, e.g. `https://reactome.org/download/current/ReactomePathways.txt`.}
}
\value{
TRUE if assertion passed
}
\description{
Stops with error message if the path is not accessible or server not
responding.
}
