% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioplanet.R
\name{fetch_bp}
\alias{fetch_bp}
\title{Get functional term data from BioPlanet}
\usage{
fetch_bp(use_cache = TRUE, on_error = c("stop", "warn"))
}
\arguments{
\item{use_cache}{Logical, if TRUE, the remote file will be cached locally.}

\item{on_error}{A character vector specifying the error handling method. It
can take values `"stop"` or `"warn"`. The default is `"stop"`. `"stop"`
will halt the function execution and throw an error, while `"warn"` will
issue a warning and return `NULL`.}
}
\value{
A list with \code{terms} and \code{mapping} tibbles.
}
\description{
Download term information (term ID and name) and gene-pathway mapping
(NCBI gene ID, gene symbol and pathway ID) from BioPlanet.
}
\examples{
bioplanet_data <- fetch_bp(on_error = "warn")
}
