% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{fetch_go_from_go}
\alias{fetch_go_from_go}
\title{Get functional term data from gene ontology}
\usage{
fetch_go_from_go(species)
}
\arguments{
\item{species}{Root name for species file under
\code{http://current.geneontology.org/annotations}. Examples are
\code{"goa_human"} for human, \code{"mgi"} for mouse or \code{"sgd"} for yeast.}
}
\value{
A list with \code{terms} and \code{mapping} tibbles.
}
\description{
Download term information (GO term ID and name) and gene-term mapping
(gene symbol and GO term ID) from gene ontology.
}
\examples{
\dontrun{
go_data <- fetch_go_from_go("sgd")
}
}
