% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{fetch_go_genes_go}
\alias{fetch_go_genes_go}
\title{Download GO term gene mapping from geneontology.org}
\usage{
fetch_go_genes_go(species)
}
\arguments{
\item{species}{Root name for species file under
\code{http://current.geneontology.org/annotations}. Examples are
\code{goa_human} for human, \code{mgi} for mouse or \code{sgd} for yeast.}
}
\value{
A tibble with gene_name, uniprot_id and term_id.
}
\description{
Download GO term gene mapping from geneontology.org
}
