% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{fetch_kegg_species}
\alias{fetch_kegg_species}
\title{Find all species available from KEGG}
\usage{
fetch_kegg_species(on_error = c("stop", "warn", "ignore"))
}
\arguments{
\item{on_error}{A character string indicating the error handling strategy:
either "stop" to halt execution, "warn" to issue a warning and return
`NULL` or "ignore" to return `NULL` without warnings. Defaults to "stop".}
}
\value{
A tibble, in which column \code{designation} contains species
  designations used in function \code{fetch_kegg}.
}
\description{
Find all species available from KEGG
}
\examples{
spe <- fetch_kegg_species(on_error = "warn")
}
