% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{fetch_kegg_species}
\alias{fetch_kegg_species}
\title{Find all species available from KEGG}
\usage{
fetch_kegg_species(on_error = c("stop", "warn"))
}
\arguments{
\item{on_error}{A character vector specifying the error handling method. It
can take values `"stop"` or `"warn"`. The default is `"stop"`. `"stop"`
will halt the function execution and throw an error, while `"warn"` will
issue a warning and return `NULL`.}
}
\value{
A tibble, in which column \code{designation} contains species
  designations used in function \code{fetch_kegg}.
}
\description{
Find all species available from KEGG
}
\examples{
spe <- fetch_kegg_species(on_error = "warn")
}
