% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{fetch_reactome}
\alias{fetch_reactome}
\title{Get functional term data from Reactome}
\usage{
fetch_reactome(species)
}
\arguments{
\item{species}{Reactome species designation, for example \code{"Homo
sapiens"} for human. Full list of available species can be found using
\code{fetch_reactome_species()}.}
}
\value{
A list with \code{terms} and \code{mapping} tibbles.
}
\description{
Download term information (pathway ID and name) and gene-pathway mapping
(Ensembl gene ID and pathway ID) from Reactome.
}
\examples{
\dontrun{
reactome_data <- fetch_reactome("Homo sapiens")
}
}
