% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{fetch_reactome}
\alias{fetch_reactome}
\title{Get functional term data from Reactome}
\usage{
fetch_reactome(
  species,
  source = c("ensembl", "api", "gene_association"),
  use_cache = TRUE,
  on_error = c("stop", "warn")
)
}
\arguments{
\item{species}{Reactome species designation, for example "Homo sapiens" for
human. Full list of available species can be found using
\code{fetch_reactome_species()}.}

\item{source}{How to download the mapping. If 'ensembl' or
'gene_association', one mapping file provided by Reactome will be
downloaded, if 'api', then Reactome API will be used. See details.}

\item{use_cache}{Logical, if TRUE, the remote file will be cached locally.}

\item{on_error}{A character vector specifying the error handling method. It
can take values `"stop"` or `"warn"`. The default is `"stop"`. `"stop"`
will halt the function execution and throw an error, while `"warn"` will
issue a warning and return `NULL`.}
}
\value{
A list with \code{terms} and \code{mapping} tibbles
}
\description{
Download term information (pathway ID and name) and gene-pathway mapping
(Ensembl gene ID or gene symbol and pathway ID) from Reactome.
}
\details{
Reactome makes mapping between Ensembl ID and pathway ID available
  in form of one downloadable file. Also, a gene association file with
  mapping between UniProt accession number, gene symbol and Reactome term is
  available.  If \code{source = "ensembl"} or \code{source =
  "gene_association"} is set, one large file will be downloaded and parsed.
  If \code{source = "api"} is set, then Reactome APIs will be interrogated
  for each pathway available. This method is considerably slower, especially
  for large genomes. However, gene association file contains far fewer
  mappings than can be extracted using API.
}
\examples{
reactome_data <- fetch_reactome("Saccharomyces cerevisiae")
}
