% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{fetch_reactome_species}
\alias{fetch_reactome_species}
\title{List of available Reactome species}
\usage{
fetch_reactome_species(on_error = c("stop", "warn", "ignore"))
}
\arguments{
\item{on_error}{A character string indicating the error handling strategy:
either "stop" to halt execution, "warn" to issue a warning and return
`NULL` or "ignore" to return `NULL` without warnings. Defaults to "stop".}
}
\value{
A tibble with species names used by Reactome.
}
\description{
List of available Reactome species
}
\examples{
re <- fetch_reactome_species(on_error = "warn")
}
