% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{fetch_wiki_pathway_genes_api}
\alias{fetch_wiki_pathway_genes_api}
\title{Download WikiPathways gene-pathway mapping using API}
\usage{
fetch_wiki_pathway_genes_api(pathways, databases = NULL, types = NULL)
}
\arguments{
\item{pathways}{A character vector with pathway names.}

\item{databases}{Names of databases to use}

\item{types}{Names of types to use}
}
\value{
A tibble with \code{term_id} and \code{gene_symbol}
}
\description{
Download WikiPathways gene-pathway mapping using API
}
