% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{fetch_wiki_species}
\alias{fetch_wiki_species}
\title{List of available WikiPathways species}
\usage{
fetch_wiki_species(on_error = c("stop", "warn", "ignore"))
}
\arguments{
\item{on_error}{A character string indicating the error handling strategy:
either "stop" to halt execution, "warn" to issue a warning and return
`NULL` or "ignore" to return `NULL` without warnings. Defaults to "stop".}
}
\value{
A character vector with species names used by WikiPathways.
}
\description{
List of available WikiPathways species
}
\examples{
spec <- fetch_wiki_species(on_error = "warn")
}
