% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{fetch_wiki_species}
\alias{fetch_wiki_species}
\title{List of available WikiPathways species}
\usage{
fetch_wiki_species(on_error = c("stop", "warn"))
}
\arguments{
\item{on_error}{A character vector specifying the error handling method. It
can take values `"stop"` or `"warn"`. The default is `"stop"`. `"stop"`
will halt the function execution and throw an error, while `"warn"` will
issue a warning and return `NULL`.}
}
\value{
A character vector with species names used by WikiPathways.
}
\description{
List of available WikiPathways species
}
\examples{
spec <- fetch_wiki_species(on_error = "warn")
}
