% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{functional_enrichment}
\alias{functional_enrichment}
\title{Fast functional enrichment}
\usage{
functional_enrichment(
  feat_all,
  feat_sel,
  term_data,
  feat2name = NULL,
  min_count = 2,
  fdr_limit = 0.05
)
}
\arguments{
\item{feat_all}{A character vector with all feature identifiers. This is the
background for enrichment.}

\item{feat_sel}{A character vector with feature identifiers in the selection.}

\item{term_data}{Functional term data, as explained in details. It can be
created using \code{prepare_for_enrichment}.}

\item{feat2name}{An optional named list to convert feature id into feature
name.}

\item{min_count}{Minimal count of features with term in the selection to be
used.}

\item{fdr_limit}{Only terms with p_adjust below this limit are returned.}
}
\value{
A tibble with enrichment results. For each term the following
  quantities are reported: N_with - number of features with this term in the
  among all features, n_with_sel - number of features with this term in the
  selection, n_expect - expected number of features with this term in the
  selection, under the null hypothesis that terms are assigned to features
  randomly, enrichment - ratio of n_with_sel / n_expect; odds_ratio - odds
  ratio for enrichment, p_value - p-value from a single hypergeometric test;
  p_adjust - p-value adjusted for multiple tests using Benjamini-Hochberg
  approach.
}
\description{
Fast functional enrichment based on hypergeometric distribution. Can be used in interactive applications.
}
\details{
Functional enrichment in a selection (e.g. significantly DE features) of
features, using hypergeometric probability. A feature can be a gene, protein,
etc. `term_data` is an object with functional term information and
feature-term assignment. It is a list of: `term2info` - a named vector term
id => term name; `term2feature` - a list term_id => vector of feature_ids;
`feature2term` - a list feature id => vector of term ids. It can be created
by `prepare_for_enrichment` function.
}
\examples{
bp <- fetch_bp()
bp_terms <- prepare_for_enrichment(bp$terms, bp$mapping, exmpl_all, feature_name = "gene_symbol")
enr <- functional_enrichment(exmpl_all, exmpl_sel, bp_terms)

}
