% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_feature_terms}
\alias{get_feature_terms}
\title{Get terms annotating a given feature.}
\usage{
get_feature_terms(term_data, feature_id)
}
\arguments{
\item{term_data}{An object class \code{fenr_terms}, created by
\code{prepare_for_enrichment}.}

\item{feature_id}{A string with a feature ID}
}
\value{
A character vector containing functional term IDs annotating given feature.
}
\description{
Get terms annotating a given feature.
}
\examples{
bp_data <- fetch_bp()
bp_terms <- prepare_for_enrichment(bp_data$terms, bp_data$mapping, feature = "gene_symbol")
trms <- get_feature_terms(bp_terms, "RECK")
}
