% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{prepare_for_enrichment}
\alias{prepare_for_enrichment}
\title{Prepare term data for enrichment analysis}
\usage{
prepare_for_enrichment(
  terms,
  mapping,
  all_features = NULL,
  feature_name = "gene_id"
)
}
\arguments{
\item{terms}{A tibble with at least two columns: \code{term_id} and
\code{term_name}. Contains information about functional term
names/descriptions.}

\item{mapping}{A tibble with at least two columns, containing mapping between
functional terms and features. One column needs to be called \code{term_id}
and the other column has a name specified by \code{feature_name} argument.
For example, if \code{mapping} contains columns \code{term_id},
\code{accession_number} and  \code{gene_symbol} then setting
\code{feature_name = "gene_symbol"} indicates that gene symbols will be
used for enrichment.}

\item{all_features}{A vector with all feature ids used as background for
enrichment. If not specified, all features found in \code{mapping} will be
used, resulting in a larger object size.}

\item{feature_name}{The name of the column in \code{mapping} tibble to be
used as feature.For example, if \code{mapping} contains columns \code{term_id},
\code{accession_number} and  \code{gene_symbol} then setting
\code{feature_name = "gene_symbol"} indicates that gene symbols will be
used for enrichment.}
}
\value{
An object class \code{fenr_terms} required by
  \code{functional_enrichment}.
}
\description{
Prepare term data downloaded with \code{fetch_*} functions for fast
enrichment analysis.
}
\details{
Takes two tibbles with functional term information (\code{terms}) and
feature mapping (\code{mapping}) and converts them into an object required by
\code{functional_enrichment} for fast analysis. Terms and mapping can be
created with database access functions in this package, for example
\code{fetch_reactome} or \code{fetch_go_from_go}.
}
\examples{
data(exmpl_all)
bp <- fetch_bp()
bp_terms <- prepare_for_enrichment(bp$terms, bp$mapping, exmpl_all, feature_name = "gene_symbol")
}
