% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{prepare_for_enrichment}
\alias{prepare_for_enrichment}
\title{Prepare term data for enrichment analysis}
\usage{
prepare_for_enrichment(
  terms,
  mapping,
  all_features = NULL,
  feature_name = "gene_id"
)
}
\arguments{
\item{terms}{Information about term names/descriptions. A tibble with columns
\code{term_id} and \code{term_name}.}

\item{mapping}{Information about term-feature mapping. A tibble with
\code{term_id} and a feature id, as identified with \code{feature_name}
argument. For example, if this tibble contains \code{gene_symbol} and
\code{term_id}, then you need to set \code{feature_name = "gene_symbol"}.}

\item{all_features}{A vector with all feature ids used as background for
enrichment. If not specified, all features found in \code{mapping} will be
used, resulting in a larger object size.}

\item{feature_name}{Which column to use from mapping table, e.g.
\code{gene_symbol} or \code{ensembl_gene_id}.}
}
\value{
An object class \code{fenr_terms} required by
  \code{functional_enrichment}.
}
\description{
Prepare term data downloaded with \code{fetch_*} functions for fast
enrichment analysis.
}
\details{
Takes two data frames with functional term information (\code{terms}) and
gene mapping (\code{mapping}) and converts them into an object required by
\code{functional_enrichment} for fast analysis. Terms and mapping can be
created with database access functions in this package, for example
\code{fetch_reactome} or \code{fetch_go_from_go}.
}
\examples{
data(exmpl_all)
bp <- fetch_bp()
bp_terms <- prepare_for_enrichment(bp$terms, bp$mapping, exmpl_all, feature_name = "gene_symbol")
}
