% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{prepare_for_enrichment}
\alias{prepare_for_enrichment}
\title{Prepare term data for enrichment analysis}
\usage{
prepare_for_enrichment(terms, mapping, all_features, feature_name = "gene_id")
}
\arguments{
\item{terms}{Information about term names/descriptions. A tibble with
columns \code{term_id} and \code{term_name}.}

\item{mapping}{Information about term-feature mapping. A tibble with
\code{term_id} and a feature id, as identified with \code{feature_name} argument. For
example, if this tibble contains \code{gene_symbol} and \code{term_id}, then
`feature_name = "gene_symbol"`.}

\item{all_features}{A vector with all feature ids (background for
enrichment).}

\item{feature_name}{Which column to use from mapping table, e.g. "gene_symbol"
or "ensembl_gene_id".}
}
\value{
An object required by `functional_enrichment`.
}
\description{
Prepare term data downloaded with \code{fetch_*} functions for fast
enrichment analysis.
}
\details{
Takes two data frames with functional term information and gene mapping and
converts them into an object required by `functional_enrichment` for fast
analysis.
}
